pyproject.toml
[project]
name = "graph-fraud"
version = "0.1.0"
description = "Educational fraud detection using graph-based methods"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "networkx>=3.2",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "python-louvain>=0.16",
    "neo4j>=6.0.3",
    "marimo>=0.17.7",
    "mlflow>=3.6.0",
    "dvc>=3.63.0",
    "dvc-s3>=3.2.2",
    "fastapi>=0.121.1",
    "uvicorn>=0.38.0",
    "prometheus-client>=0.23.1",
    "great-expectations>=1.9.0",
    "apache-airflow>=3.1.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "boto3>=1.40.61",
    "optuna>=4.6.0",
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "seaborn>=0.13.2",
    "plotly>=6.4.0",
    "pyvis>=0.3.2",
    "black>=24.0.0",
    "ruff>=0.6.0",
]
